+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to `wp-log` will be documented in this file
 
 ## Unreleased
 
+- use regex to split up log items instead of using PHP_EOL
+- make timestamp of log entry bold in log interface
+
 ## 0.2.1
 
 - fix slug on limit links in interface